+++ sdk/search/Azure.Search.Documents/src/Spatial/SpatialProxyFactory.cs
@@ -8,7 +8,7 @@ using System.Reflection;
 namespace Azure.Search.Documents
 {
     /// <summary>
-    /// Creates <see cref="GeometryProxy"/> instances from unknown objects.
+    /// Creates <see cref="GeographyProxy"/> instances from unknown objects.
     /// </summary>
     internal static class SpatialProxyFactory
     {
@@ -17,13 +17,12 @@ namespace Azure.Search.Documents
         // TODO: When GeoJson is built into Azure.Core, change these proxies to adapter for GeoJson types.
         // https://github.com/Azure/azure-sdk-for-net/issues/13319
 
-        private static readonly IReadOnlyDictionary<string, Func<object, GeometryProxy>> s_types =
-            new Dictionary<string, Func<object, GeometryProxy>>(StringComparer.OrdinalIgnoreCase)
+        private static readonly IReadOnlyDictionary<string, Func<object, GeographyProxy>> s_types =
+            new Dictionary<string, Func<object, GeographyProxy>>(StringComparer.OrdinalIgnoreCase)
             {
-                ["Microsoft.Spatial.GeometryLineString"] = value => new GeometryLineStringProxy(value),
-                ["Microsoft.Spatial.GeometryPoint"] = value => new GeometryPointProxy(value),
-                ["Microsoft.Spatial.GeometryPolygon"] = value => new GeometryPolygonProxy(value),
-                ["Microsoft.Spatial.GeometryPosition"] = value => new GeometryPositionProxy(value),
+                ["Microsoft.Spatial.GeographyLineString"] = value => new GeographyLineStringProxy(value),
+                ["Microsoft.Spatial.GeographyPoint"] = value => new GeographyPointProxy(value),
+                ["Microsoft.Spatial.GeographyPolygon"] = value => new GeographyPolygonProxy(value),
             };
 
         /// <summary>
@@ -42,20 +41,20 @@ namespace Azure.Search.Documents
         }
 
         /// <summary>
-        /// Creates a <see cref="GeometryProxy"/> from the given <paramref name="value"/> if supported.
+        /// Creates a <see cref="GeographyProxy"/> from the given <paramref name="value"/> if supported.
         /// <seealso cref="CanCreate(Type)"/>
         /// </summary>
         /// <param name="value">The value to proxy.</param>
-        /// <returns>A <see cref="GeometryProxy"/> from the given <paramref name="value"/> if supported, or null if <paramref name="value"/> is null.</returns>
+        /// <returns>A <see cref="GeographyProxy"/> from the given <paramref name="value"/> if supported, or null if <paramref name="value"/> is null.</returns>
         /// <exception cref="NotSupportedException">The <paramref name="value"/> type is not supported.</exception>
-        public static GeometryProxy Create(object value)
+        public static GeographyProxy Create(object value)
         {
             if (value is null)
             {
                 return null;
             }
 
-            if (TryCreate(value, out GeometryProxy proxy))
+            if (TryCreate(value, out GeographyProxy proxy))
             {
                 return proxy;
             }
@@ -64,18 +63,18 @@ namespace Azure.Search.Documents
         }
 
         /// <summary>
-        /// Attempts to creates a <see cref="GeometryProxy"/> from the given <paramref name="value"/> if supported.
+        /// Attempts to creates a <see cref="GeographyProxy"/> from the given <paramref name="value"/> if supported.
         /// <seealso cref="CanCreate(Type)"/>
         /// </summary>
         /// <param name="value">The value to proxy.</param>
         /// <param name="proxy">The proxied value if supported.</param>
         /// <returns>True if the <paramref name="value"/> could be proxied; otherwise, false.</returns>
-        internal static bool TryCreate(object value, out GeometryProxy proxy)
+        internal static bool TryCreate(object value, out GeographyProxy proxy)
         {
             if (value is { })
             {
                 Type type = value.GetType();
-                if (TryGetFactory(type, out Func<object, GeometryProxy> factory))
+                if (TryGetFactory(type, out Func<object, GeographyProxy> factory))
                 {
                     proxy = factory(value);
                     return true;
@@ -86,7 +85,7 @@ namespace Azure.Search.Documents
             return false;
         }
 
-        private static bool TryGetFactory(Type type, out Func<object, GeometryProxy> factory)
+        private static bool TryGetFactory(Type type, out Func<object, GeographyProxy> factory)
         {
             // TODO: Should we compare the public key token as well, or is that too restrictive or unnecessarily slow?
             AssemblyName assemblyName = type.Assembly.GetName();